hermes/pyproject.toml
[project]
name = "hermes"
version = "0.1.0"
description = "Personal sales assistant daemon for the head of sales at Vertical Impression."
requires-python = ">=3.11"
dependencies = [
    "google-auth>=2.30",
    "google-auth-oauthlib>=1.2",
    "google-api-python-client>=2.140",
    "slack-sdk>=3.27",
    "python-telegram-bot>=21.5",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "tenacity>=9.0",
]

[project.optional-dependencies]
research = ["playwright>=1.47"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
